backend/requirements.txt
# AI Studio 백엔드 서버 요구사항

flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
